+++ packages/boot/src/components/root/single-page.tsx
@@ -16,6 +16,7 @@ import { EditorSnackbars } from '@wordpress/editor';
  * Internal dependencies
  */
 import SavePanel from '../save-panel';
+import CanvasRenderer from '../canvas-renderer';
 import { unlock } from '../../lock-unlock';
 import type { CanvasData } from '../../store/types';
 import './style.scss';
@@ -34,6 +35,8 @@ export default function RootSinglePage() {
 		| CanvasData
 		| null
 		| undefined;
+	const routeContentModule = ( currentMatch?.loaderData as any )
+		?.routeContentModule as string | undefined;
 	const isFullScreen = canvas && ! canvas.isPreview;
 
 	return (
@@ -54,6 +57,15 @@ export default function RootSinglePage() {
 						>
 							<Outlet />
 						</ThemeProvider>
+						{ /* Render Canvas in Root to prevent remounting on route changes */ }
+						{ ( canvas || canvas === null ) && (
+							<div className="boot-layout__canvas">
+								<CanvasRenderer
+									canvas={ canvas }
+									routeContentModule={ routeContentModule }
+								/>
+							</div>
+						) }
 					</div>
 				</div>
 			</ThemeProvider>