+++ packages/fields/src/types.ts
@@ -112,6 +112,10 @@ export type PostWithPermissions = Post & {
 	};
 };
 
+interface EditorSupport {
+	notes?: boolean;
+}
+
 export interface PostType {
 	slug: string;
 	viewable: boolean;
@@ -122,7 +126,7 @@ export interface PostType {
 		author?: string;
 		thumbnail?: string;
 		comments?: string;
-		editor?: boolean;
+		editor?: boolean | [ EditorSupport ];
 		trackbacks?: boolean;
 	};
 }